+++ umami/plot_input_variables.py
@@ -1,7 +1,5 @@
-#!/usr/bin/env python
 """
-This script plots the given input variables of the given files and
-also a comparison.
+This script plots the given input variables of the given files and also a comparison.
 """
 
 import argparse