+++ umami/preprocessing.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 """Execution script to run preprocessing steps."""
 import argparse