+++ umami/tests/unit/plotting_tools/test_train_plot_functions.py
@@ -1,8 +1,4 @@
-#!/usr/bin/env python
-
-"""
-Unit test script for the train plot functions.
-"""
+"""Unit test script for the train plot functions."""
 
 import os
 import tempfile