+++ umami/tests/unit/train_tools/test_NN_tools.py
@@ -1,9 +1,4 @@
-#!/usr/bin/env python
-
-"""
-Unit test script for the NN_tools functions
-"""
-
+"""Unit test script for the NN_tools functions."""
 import json
 import os
 import tempfile